pyproject.toml
[project]
name = "mind-to-project"
version = "0.1.0"
description = "CLI tool to turn raw brain-dumps into clean project descriptions via AI pipeline"
readme = "README.md"
authors = [
    { name = "Bernhard Raml", email = "bernhard.raml@geo.tuwien.ac.at" }
]
requires-python = ">=3.12"
dependencies = [
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
]

[dependency-groups]
dev = [
    "pyright>=1.1.408",
    "pytest>=7.0.0",
    "pytest-mock>=3.0.0",
    "ruff>=0.15.9",
]

[project.scripts]
mind-to-project = "mind_to_project:main"

[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"
